add transactions list for stablecoins
