+++ src/ducks/Stablecoins/StablecoinSelector/StablecoinSelector.module.scss
@@ -1,0 +1,77 @@
+@import '~@santiment-network/ui/mixins';
+
+.projectIcon {
+  border: 1px solid var(--porcelain);
+  box-sizing: border-box;
+  box-shadow: 0 4px 8px rgba(24, 27, 43, 0.03);
+  width: 40px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 50%;
+}
+
+.project {
+  margin-left: 16px;
+
+  @include responsive('tablet') {
+    max-width: 278px;
+  }
+
+  &__top {
+    display: flex;
+    align-items: center;
+    fill: var(--mirage);
+  }
+
+  &__name {
+    margin: 0 0 4px;
+    color: var(--rhino);
+    font-size: 24px !important;
+    line-height: 32px !important;
+
+    @include text('h4');
+  }
+
+  &__arrows {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    margin: 0 0 0 12px;
+    height: 8px;
+  }
+
+  &__arrow {
+    fill: var(--waterloo);
+  }
+
+  &__description {
+    @include text('body-3');
+
+    color: var(--waterloo);
+    max-width: 450px;
+    margin-bottom: 16px;
+
+    @include responsive('laptop') {
+      max-width: 400px;
+    }
+  }
+}
+
+.selector {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+
+  &:hover {
+    & .project__arrow {
+      fill: var(--jungle-green-hover);
+    }
+
+    & .project__name {
+      color: var(--jungle-green-hover);
+    }
+  }
+}
