+++ src/ducks/Stablecoins/StablecoinsTransactions/StablecoinsTransactions.module.scss
@@ -1,0 +1,14 @@
+@import '~@santiment-network/ui/mixins';
+
+.container {
+  box-sizing: border-box;
+  border-radius: $border-radius;
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  margin-bottom: 24px;
+  padding: 16px;
+  background: var(--white);
+}
