+++ src/pages/Detailed/transactionsInfo/DetailedTransactionsTable.module.scss
@@ -26,6 +26,10 @@
       }
     }
 
+    .rt-noData {
+      background: none;
+    }
+
     .rt-th {
       text-align: left;
       padding: 7px 0;
